requirements.txt
ccxt
dash
dash-bootstrap-components
matplotlib
plotly
pandas
python-dotenv
quandl
scipy
statsmodels
yfinance
numpy